removing the useless part (war), renaming packages and correcting for the transaction mode
