--- idsteptodeletetest/src/main/java/fr/abes/idsteptodeletetest/IdsteptodeletetestApplication.java
@@ -1,40 +1,0 @@
-package fr.abes.idsteptodeletetest;
-
-import org.springframework.batch.core.explore.JobExplorer;
-import org.springframework.batch.core.launch.JobLauncher;
-import org.springframework.batch.core.launch.support.CommandLineJobRunner;
-import org.springframework.batch.core.repository.JobRepository;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.batch.BatchAutoConfiguration;
-import org.springframework.boot.autoconfigure.batch.JobLauncherCommandLineRunner;
-import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
-import org.springframework.boot.autoconfigure.jdbc.DataSourceTransactionManagerAutoConfiguration;
-import org.springframework.boot.autoconfigure.orm.jpa.HibernateJpaAutoConfiguration;
-import org.springframework.context.ApplicationContext;
-import org.springframework.transaction.annotation.EnableTransactionManagement;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-
-
-@SpringBootApplication(//scanBasePackages = {"fr.abes.idsteptodeletetest.JpaBatchConfigurer"},
-        //scanBasePackageClasses = fr.abes.idsteptodeletetest.JpaBatchConfigurer.class,
-        exclude = { DataSourceAutoConfiguration.class,
-                HibernateJpaAutoConfiguration.class,
-                DataSourceTransactionManagerAutoConfiguration.class,
-                BatchAutoConfiguration.class //si on ne met pas ça, il déroule tous les beans job avec resultat NOOP
-        }
-)
-@EnableTransactionManagement
-public class IdsteptodeletetestApplication {
-
-    private Logger logger = LogManager.getLogger(IdsteptodeletetestApplication.class);
-
-    public static void main(String[] args) {
-        ApplicationContext ctx = SpringApplication.run(IdsteptodeletetestApplication.class, args);
-
-    }
-}
-
-
-
