--- idsteptodeletetest/src/main/java/fr/abes/idsteptodeletetest/MessageUsage.java
@@ -1,4 +1,0 @@
-package fr.abes.idsteptodeletetest;
-
-public class MessageUsage {
-}
