+++ src/main/java/fr/abes/ongoingthesestodelete/MessageUsage.java
@@ -1,0 +1,4 @@
+package fr.abes.ongoingthesestodelete;
+
+public class MessageUsage {
+}
