--- src/main/java/fr/abes/testrestcontrollerinterface/JarExecutor.java
@@ -1,71 +1,0 @@
-package fr.abes.testrestcontrollerinterface;
-
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.util.ArrayList;
-import java.util.List;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-public class JarExecutor {
-
-    private BufferedReader error;
-    private BufferedReader op;
-    private int exitVal;
-    private static final Logger logger = LoggerFactory
-            .getLogger(JarExecutor.class);
-
-    public void executeJar(String jarFilePath, List<String> args) throws Exception {
-        // Create run arguments for the
-
-        final List<String> actualArgs = new ArrayList<String>();
-        actualArgs.add(0, "java");
-        actualArgs.add(1, "-jar");
-        actualArgs.add(2, jarFilePath);
-        actualArgs.addAll(args);
-        try {
-            final Runtime re = Runtime.getRuntime();
-            //final Process command = re.exec(cmdString, args.toArray(new String[0]));
-            final Process process = re.exec(actualArgs.toArray(new String[0]));
-            logger.info("command = " + process);
-            this.error = new BufferedReader(new InputStreamReader(process.getErrorStream()));
-            this.op = new BufferedReader(new InputStreamReader(process.getInputStream()));
-            // Wait for the application to Finish
-            while ((op.readLine()) != null) {}
-            process.waitFor();
-            //command.waitFor();
-            this.exitVal = process.exitValue();
-            if (this.exitVal != 0) {
-                throw new IOException("Failed to execure jar, " + this.getExecutionLog());
-            }
-
-        } catch (final IOException | InterruptedException e) {
-            throw new Exception(e);
-        }
-    }
-
-    public String getExecutionLog() {
-        String error = "";
-        String line;
-        try {
-            while((line = this.error.readLine()) != null) {
-                error = error + "\n" + line;
-            }
-        } catch (final IOException e) {
-        }
-        String output = "";
-        try {
-            while((line = this.op.readLine()) != null) {
-                output = output + "\n" + line;
-            }
-        } catch (final IOException e) {
-        }
-        try {
-            this.error.close();
-            this.op.close();
-        } catch (final IOException e) {
-        }
-        return "exitVal: " + this.exitVal + ", error: " + error + ", output: " + output;
-    }
-}
