--- src/main/java/fr/abes/testrestcontrollerinterface/ServletInitializer.java
@@ -1,14 +1,0 @@
-package fr.abes.testrestcontrollerinterface;
-
-import org.springframework.boot.builder.SpringApplicationBuilder;
-import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
-
-
-public class ServletInitializer extends SpringBootServletInitializer {
-
-    @Override
-    protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
-        return application.sources(TestrestcontrollerinterfaceApplication.class);
-    }
-
-}
