--- src/main/java/fr/abes/testrestcontrollerinterface/TestrestcontrollerinterfaceApplication.java
@@ -1,18 +1,0 @@
-package fr.abes.testrestcontrollerinterface;
-
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.ApplicationContext;
-
-
-
-@SpringBootApplication()
-
-public class TestrestcontrollerinterfaceApplication {
-
-    public static void main(String[] args) {
-        ApplicationContext ctx = SpringApplication.run(TestrestcontrollerinterfaceApplication.class, args);
-    }
-}
-
