+++ src/main/resources/admin_suppressionmasse.txt
@@ -1,0 +1,15 @@
+100022
+100023
+100024
+100032
+100041
+100046
+100048
+100049
+100050
+100051
+100055
+100058
+100059
+100060
+100061
