--- src/main/resources/data.csv
@@ -1,2 +1,0 @@
-121748,
-121983,
