+++ src/main/resources/sh/admin_suppressionmasse.sh
